// src/constants/enum.ts
export enum Input {
	DATE = "DATE",
	EMAIL = "EMAIL",
	NUMERIC = "NUMERIC",
	PASSWORD = "PASSWORD",
	SWITCH = "SWITCH",
	SELECT = "SELECT",
	SEARCHABLE_SELECT = "SEARCHABLE_SELECT",
	TEXT = "TEXT",
	PHONE = "PHONE",
	FILTER_TEXT = "FILTER_TEXT",
	ADDRESS_SEZNAM = "ADDRESS-S",
	ADDRESS_GOOGLE = "ADDRESS-G",
	CUSTOMER = "CUSTOMER",
}

export enum Method {
	POST = "POST",
	GET = "GET",
	PUT = "PUT",
	DELETE = "DELETE",
	PATCH = "PATCH",
}